Move minio outside kuttl tests
MinIO was being started as a part of the rclone e2e, but it will also be
used for the upcoming restic mover as well. This change adds a step to
CI to start minio when the cluster is brought up so that it can be
shared.

This also changes the name of the Scribe Helm deployment so that we can
catch problems with fixed names in the future :)

Signed-off-by: John Strunk <jstrunk@redhat.com>