Make restic cache volume configurable
This adds the ability to configure the capacity, StorageClass, and
AccessMode of the cache volume used by Restic. This is needed because
the backup metadata must fit within the cache volume for a backup,
restore, or prune to succeed. For volumes with a large number of files,
the default 1Gi is insufficient.

Signed-off-by: John Strunk <jstrunk@redhat.com>