Race in set-replication
Describe what this PR does

Issue:
When set-replication performs the manual sync, it waits for the sync to finish
on the source (trigger status == spec), then it immediately attempts to retrieve
the latestImage from the ReplicationDestination. There is no guarantee that
latestImage has been updated yet, even though we know the sync has completed from
the PoV of the source.

Solution:
1. Check destination and remember what we find
2. Do the manual sync on the source
3. Poll the destination until it's different than what we saw in step 1 (or timeout)

Is there anything that requires special attention?

Related issues:
Fixes #14

Signed-off-by: Vinayakswami Hariharmath <vharihar@redhat.com>