Increase e2e timeout when waiting for snapshots
This increases the timeout for test steps that use volumes created from
snapshots. This is necessary because cloud providers can take a while
for Snapshots to become ready to use. It also increases the sync
interval to ensure we don't start a new sync while waiting for snapshots
from the previous.

Signed-off-by: John Strunk <jstrunk@redhat.com>