rsync-tls: Give time for rsync to complete
Periodically, the source seems to think the replication failed, while
the destination registers success. This seems to be due to the
destination shutting down before the rsync/stunnel connection properly
terminates. This causes the source to believe the "shutdown" operation
failed, and it restarts the replication.
This change adds a sleep on the destination to give the rsync & stunnel
time to close the connection before terminating stunnel.
It also adds a corresponding sleep on the source to ensure the source
replication can't restart before the destination has terminated.
Handling the race in this manner seems like a poor solution, but I'm at
a loss for a better approach.

Signed-off-by: John Strunk <jstrunk@redhat.com>