Added diskrsync to rsync ssh mover
Added diskrsync https://github.com/dop251/diskrsync
for block devices to the ssh mover. This is to support
syncing KubeVirt virtual machines, which primarily use
block devices for the disk images.

diskrsync does not support TLS so I could not add it to
the tls mover. If this is important, we will have to open
a PR on diskrsync to add support for TLS.

Signed-off-by: Alexander Wels <awels@redhat.com>