restic restore only user.* xattrs
- see issue: https://github.com/restic/restic/issues/5089

This issue occured only in OpenShift when using restic in
privileged mode. With restic v0.17.1 and later, restoring xattrs
would cause the restore to fail if UID=0 if there were any
issues restoring xattrs.

When using privileged, we run as uid 0 and on some kube clusters
such as OpenShift, we were unable to restore things like
security.selinux settings. This had never worked, but previous
verions of restic woudl not consider it an error.

Workaround right now is to stop trying to restore security,
system and trusted namespaced xattrs.

Easiest just to use the --include-xattrs flag to specify that
we only want to attempt to restore user.* xattrs.

Signed-off-by: Tesshu Flower <tflower@redhat.com>